task/MAR10001-1545: - Added support for configured price rounding - Added support for configured price precisions - Renamed marello_productprice to marello_pricing alias - Added rounding fixes for Prices and Order related entities

+++ src/Marello/Bundle/TaxBundle/Provider/OrderItemRowTotalsProvider.php
@@ -63,7 +63,7 @@ class OrderItemRowTotalsProvider extends AbstractOrderItemFormChangesProvider
                     } else {
                         $rate = 0;
                     }
-                    $amount = (double)$itemResult['price'][$identifier]['value'] * (float)$item['quantity'];
+                    $amount = (float)$itemResult['price'][$identifier]['value'] * (float)$item['quantity'];
                     $taxTotals = $this->taxCalculator->calculate($amount, $rate);
                     $itemResult['row_totals'][$identifier][$rowId] = $taxTotals->jsonSerialize();
                 }
